+++ go/api/glide_cluster_client.go
@@ -836,16 +836,16 @@ func (client *GlideClusterClient) ConfigSetWithOptions(
 // [valkey.io]: https://valkey.io/commands/config-get/
 func (client *GlideClusterClient) ConfigGet(
 	parameters []string,
-) (ClusterValue[interface{}], error) {
+) (map[string]string, error) {
 	res, err := client.executeCommand(C.ConfigGet, parameters)
 	if err != nil {
-		return createEmptyClusterValue[interface{}](), err
+		return nil, err
 	}
-	data, err := handleInterfaceResponse(res)
+	data, err := handleStringToStringMapResponse(res)
 	if err != nil {
-		return createEmptyClusterValue[interface{}](), err
+		return nil, err
 	}
-	return createClusterValue[interface{}](data), nil
+	return data, nil
 }
 
 // Get the values of configuration parameters.
@@ -864,16 +864,23 @@ func (client *GlideClusterClient) ConfigGet(
 // [valkey.io]: https://valkey.io/commands/config-get/
 func (client *GlideClusterClient) ConfigGetWithOptions(
 	parameters []string, opts options.RouteOption,
-) (ClusterValue[interface{}], error) {
+) (ClusterValue[map[string]string], error) {
 	res, err := client.executeCommandWithRoute(C.ConfigGet, parameters, opts.Route)
 	if err != nil {
-		return createEmptyClusterValue[interface{}](), err
+		return createEmptyClusterValue[map[string]string](), err
 	}
-	data, err := handleInterfaceResponse(res)
+	if opts.Route == nil || !opts.Route.IsMultiNode() {
+		data, err := handleStringToStringMapResponse(res)
+		if err != nil {
+			return createEmptyClusterValue[map[string]string](), err
+		}
+		return createClusterSingleValue[map[string]string](data), nil
+	}
+	data, err := handleMapOfStringMapResponse(res)
 	if err != nil {
-		return createEmptyClusterValue[interface{}](), err
+		return createEmptyClusterValue[map[string]string](), err
 	}
-	return createClusterValue[interface{}](data), nil
+	return createClusterMultiValue[map[string]string](data), nil
 }
 
 // Set the name of the current connection.
@@ -1585,4 +1592,53 @@ func (client *GlideClusterClient) FunctionStatsWithRoute(
 
 	// For multi-node routes, return the map of node addresses to FunctionStatsResult
 	return createClusterMultiValue[FunctionStatsResult](stats), nil
+}
+
+// Deletes a library and all its functions.
+// The command will be routed to all primary nodes.
+//
+// Since:
+//
+//	Valkey 7.0 and above.
+//
+// See [valkey.io] for details.
+//
+// Parameters:
+//
+//	libName - The library name to delete.
+//
+// Return value:
+//
+//	"OK" if the library exists, otherwise an error is thrown.
+//
+// [valkey.io]: https://valkey.io/commands/function-delete/
+func (client *GlideClusterClient) FunctionDelete(libName string) (string, error) {
+	return client.FunctionDeleteWithRoute(libName, options.RouteOption{})
+}
+
+// Deletes a library and all its functions.
+//
+// Since:
+//
+//	Valkey 7.0 and above.
+//
+// See [valkey.io] for details.
+//
+// Parameters:
+//
+//	libName - The library name to delete.
+//	route - Specifies the routing configuration for the command. The client will route the
+//	    command to the nodes defined by `route`.
+//
+// Return value:
+//
+//	"OK" if the library exists, otherwise an error is thrown.
+//
+// [valkey.io]: https://valkey.io/commands/function-delete/
+func (client *GlideClusterClient) FunctionDeleteWithRoute(libName string, route options.RouteOption) (string, error) {
+	result, err := client.executeCommandWithRoute(C.FunctionDelete, []string{libName}, route.Route)
+	if err != nil {
+		return DefaultStringResponse, err
+	}
+	return handleStringResponse(result)
 }